+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - added: (Zano) BTCx token
 - added: (Zcash) ZIP-321 URI support
+- changed: (Polygon) Decrease gas polling frequency to 1 minute
+- fixed: (Polygon) Use `maxFee` instead of `maxPriorityFee` for gas station parsing
 
 ## 4.65.1 (2025-11-28)